requirements.txt
snowflake-connector-python
snowflake-snowpark-python
python-dotenv
streamlit
pandas